+++ tensorflow/core/kernels/segment_reduction_ops.cc
@@ -466,7 +466,7 @@ class SparseSegmentGradOpBase : public OpKernel {
     for (int64 i = 0; i < N; ++i) {
       scaling[segment_vec(i)] += 1;
     }
-    for (int i = 0; i < scaling.size(); ++i) {
+    for (size_t i = 0; i < scaling.size(); ++i) {
       if (is_sqrtn_) {
         scaling[i] = 1.0 / sqrt(std::max(scaling[i], 1.0));
       } else {
